requirements.txt
google-adk>=1.2.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
pandas>=2.2.0
matplotlib>=3.9.0
plotly>=5.24.0
kaleido>=0.2.1
numpy>=1.26.0
scikit-learn>=1.5.0
scipy>=1.13.0
lifelines>=0.29.0
networkx
mlxtend>=0.23.0
statsmodels>=0.14.0
